requirements.txt
# Core dependencies
openai>=1.33.0
pydantic>=2.6.1
pydantic-settings>=2.1.0

# FastAPI and server
fastapi>=0.109.2
uvicorn>=0.27.1

# Environment and config
python-dotenv>=1.0.0
pyyaml>=6.0.1

# HTTP client
aiohttp>=3.9.3

# Type hints
typing-extensions>=4.9.0

# Development tools
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2